+++ lib/native_ui.dart
@@ -4,3 +4,4 @@ export 'src/button.dart';
 export 'src/ios/button.dart';
 export 'src/platform_switcher.dart';
 export 'src/conditional_builder.dart';
+export 'src/ios/form.dart';